add incremental tests
